Split out test_direct_links
